requirements.txt
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
dj-database-url==0.5.0
Django==2.2.3
django-cors-headers==3.0.1
django-heroku==0.3.1
django-mysql==3.2.0
djangorestframework==3.10.1
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.8
joblib==0.13.2
kiwisolver==1.1.0
matplotlib==2.2.4
nltk==3.4.3
numpy==1.16.4
pandas==0.25.0
Pillow==6.1.0
psycopg2==2.8.3
PyJWT==1.7.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
six==1.12.0
sqlparse==0.3.0
textblob==0.15.3
unicode==2.7
Unidecode==1.1.1
urllib3==1.25.3
vincent==0.4.4
virtualenv==15.1.0
whitenoise==4.1.3
wordcloud==1.5.0
